LPS-54106 Moves to taglib to reuse it
